improve dispatch of triangulation backend
