+++ napari/layers/shapes/_accelerated_triangulate_numba.py
